match lang changes and start EMI item application

+++ src/main/java/com/simibubi/create/compat/emi/CreateEmiRecipe.java
@@ -152,7 +152,7 @@ public abstract class CreateEmiRecipe<T extends Recipe<?>> implements EmiRecipe 
 	}
 
 	public static EmiStack fluidStack(FluidStack stack) {
-		return EmiStack.of(stack.getType(), (int) stack.getAmount());
+		return EmiStack.of(stack.getType(), stack.getAmount());
 	}
 
 	public static TextureWidget addTexture(WidgetHolder widgets, AllGuiTextures texture, int x, int y) {
